Moved the complex toUTF8Bytes() implementation that was having issues into the new depricated method 'toUTF8BytesOld()'.
Replaced the implementation in toUTF8Bytes() into a *slightly* more functional implementation (which now works with the ascii and the latinSmallLetterZWithDescender, but not the damn musical note). This implementation casts the codePoint into a char, then into a String, then uses the java String method to convert into a UTF-8 byte array.

I think the problem with this method and the musical note codepoint has to do with the signed value of the musical note, and I forget from class how to fix that....